requirements.txt
tgalice
django-user-agents
